refactor signup page
