--- client/components/SignupForm.vue
@@ -1,171 +1,0 @@
-<template>
-  <vs-row vs-align="center" vs-w="8" vs-lg="8" vs-sm="8" vs-xs="10">
-    <h2 class="form-title">Projectに登録</h2>
-    <div class="input-container">
-      <validation-observer ref="observer" v-slot="{ invalid }" tag="form">
-        <div class="form_height">
-          <vs-col>
-            <validation-provider
-              v-slot="{ errors }"
-              rules="required|min:5|max:20"
-              name="ユーザー名"
-            >
-              <vs-input v-model="form.username" label="ユーザ名" size="large" />
-              <InputError :errors="errors" />
-            </validation-provider>
-          </vs-col>
-        </div>
-        <vs-col>
-          <validation-provider
-            v-slot="{ errors }"
-            rules="required|email"
-            name="メールアドレス"
-          >
-            <vs-input
-              v-model="form.email"
-              label="メールアドレス"
-              size="large"
-            />
-            <InputError :errors="errors" />
-          </validation-provider>
-        </vs-col>
-        <vs-col>
-          <validation-provider
-            v-slot="{ errors }"
-            rules="required|min:6|max:20"
-            name="パスワード"
-          >
-            <vs-input
-              v-model="form.password"
-              type="password"
-              label="パスワード"
-              size="large"
-            />
-            <InputError :errors="errors" />
-          </validation-provider>
-        </vs-col>
-        <vs-col vs-type="flex" vs-justify="start" vs-w="12">
-          <validation-provider rules="required:true" name="同意">
-            <vs-checkbox v-model="form.isAgreed" icon="check" vs-value="ok"
-              >利用規約に同意します</vs-checkbox
-            >
-          </validation-provider>
-        </vs-col>
-        <vs-col vs-type="flex" vs-justify="start" vs-w="12">
-          <vs-button
-            size="large"
-            color="rgb(25, 150, 254)"
-            type="filled"
-            :disabled="invalid"
-            @click="onSubmit"
-            >アカウントを作成
-          </vs-button>
-        </vs-col>
-      </validation-observer>
-      <vs-col vs-type="flex" vs-justify="center" vs-align="center" vs-w="12">
-        <vs-divider>または</vs-divider>
-      </vs-col>
-
-      <vs-col
-        vs-type="flex"
-        vs-justify="center"
-        vs-align="center"
-        vs-w="12"
-        class="vs-col-google"
-      >
-        <vs-button
-          size="large"
-          color="danger"
-          type="filled"
-          vs-w="12"
-          @click="onClickGoogleButton"
-        >
-          Googleアカウントでログイン
-        </vs-button>
-      </vs-col>
-    </div>
-  </vs-row>
-</template>
-
-<script lang="ts">
-import Vue from 'vue'
-import InputError from '@/components/InputError.vue'
-import { authStore } from '@/store'
-import { buildApiUrl } from '@/store/utils'
-
-type Data = {
-  form: {
-    username: string
-    email: string
-    password: string
-    isAgreed: string | undefined
-  }
-}
-export default Vue.extend({
-  components: {
-    InputError
-  },
-  data(): Data {
-    return {
-      form: {
-        username: '',
-        email: '',
-        password: '',
-        isAgreed: undefined
-      }
-    }
-  },
-  methods: {
-    async onSubmit() {
-      this.$vs.loading()
-      const { error, messages } = await authStore.signup(this.form)
-      this.$vs.loading.close()
-
-      if (!error) {
-        this.form.username = ''
-        this.form.email = ''
-        this.form.password = ''
-
-        this.$router.push('/auth/mailsend')
-      } else if (error && messages) {
-        this.notify({
-          messages,
-          color: 'warning'
-        })
-      }
-    },
-    onClickGoogleButton() {
-      const apiUrl = buildApiUrl()
-      window.location.href = `${apiUrl}/api/auth/google`
-    }
-  }
-})
-</script>
-
-<style lang="scss" scoped>
-.form-title {
-  font-size: calc(20px + 1vw);
-  color: #707070;
-}
-.input-container {
-  margin-top: 25px;
-  .vs-col {
-    margin-bottom: 20px;
-    color: #777777;
-    font-family: HiraginoSans-W5;
-    .vs-input {
-      width: 100%;
-    }
-    .vs-button-text {
-      align-content: center;
-    }
-    .vs-button {
-      margin-top: 20px;
-      width: 100%;
-    }
-  }
-  .vs-col-google {
-    margin-top: -15px;
-  }
-}
-</style>
